--- src/main/java/net/gunivers/gunibot/datas/serialize/Loadable.java
@@ -1,6 +1,0 @@
-package net.gunivers.gunibot.datas.serialize;
-
-public interface Loadable {
-
-	public void load(Serializer serializer);
-}
