--- src/main/java/net/gunivers/gunibot/datas/serialize/Restorable.java
@@ -1,5 +1,0 @@
-package net.gunivers.gunibot.datas.serialize;
-
-public interface Restorable extends Saveable, Loadable {
-
-}
