--- src/main/java/net/gunivers/gunibot/datas/serialize/Saveable.java
@@ -1,6 +1,0 @@
-package net.gunivers.gunibot.datas.serialize;
-
-public interface Saveable {
-
-	public Serializer save();
-}
